Move test for getPluginApimlConfigs into Services.unit.test.ts
