+++ packages/core/__tests__/apiml/__unit__/Services.unit.test.ts
@@ -11,7 +11,9 @@
 
 import { Services } from "../../../src/apiml/Services";
 import { ZosmfRestClient } from "../../../src/rest/ZosmfRestClient";
-import { ConfigConstants, ImperativeError, RestConstants, Session } from "@zowe/imperative";
+import { ConfigConstants, ImperativeConfig, ImperativeError, PluginManagementFacility,
+         RestConstants, Session
+} from "@zowe/imperative";
 import { IApimlProfileInfo } from "../../../src/apiml/doc/IApimlProfileInfo";
 import * as JSONC from "comment-json";
 
@@ -24,8 +26,64 @@ describe("APIML Services unit tests", () => {
     });
 
     describe("getPluginApimlConfigs", () => {
-        it("should be tested", () => {
-            expect(true).toBe(false);
+        it("should throw an error if Imperative.init has NOT been called", () => {
+            let caughtError: Error;
+            try {
+                Services.getPluginApimlConfigs();
+            } catch(error) {
+                caughtError = error;
+            }
+            expect(caughtError).toBeInstanceOf(ImperativeError);
+            expect(caughtError.message).toContain("Imperative.init() must be called before getPluginApimlConfigs()");
+        });
+
+        it("should form apiml configs for Zowe-CLI and plugins", async () => {
+            // get an imperative config from a test file
+            const loadedConfigMock = require("./cliImpConfigMock.json");
+
+            // getPluginApimlConfigs calls ImperativeConfig.instance functions
+            // that are getters of properties, so mock the getters.
+            Object.defineProperty(ImperativeConfig.instance, "loadedConfig", {
+                configurable: true,
+                get: jest.fn(() => loadedConfigMock)
+            });
+            Object.defineProperty(ImperativeConfig.instance, "hostPackageName", {
+                configurable: true,
+                get: jest.fn(() => "@zowe/cli")
+            });
+
+            // get all plugin config props from a test file
+            const allPluginCfgPropsMock = require("./allPluginCfgPropsMock.json");
+
+            // getPluginApimlConfigs calls PluginManagementFacility.instance functions
+            // that are getters of properties, so mock the getters.
+            Object.defineProperty(PluginManagementFacility.instance, "allPluginCfgProps", {
+                configurable: true,
+                get: jest.fn(() => allPluginCfgPropsMock)
+            });
+
+            // here's the thing we want to test
+            const apimlConfigs = Services.getPluginApimlConfigs();
+
+            // debug: console.log("apimlConfigs:\n" + JSON.stringify(apimlConfigs, null, 2));
+            const numOfApimlConfigs = 4;
+            const zosmfInx = 0;
+            const endvInx1 = 1;
+            const endvInx2 = 2;
+            const jckInx = 3;
+
+            expect(apimlConfigs.length).toBe(numOfApimlConfigs);
+
+            // we get zosmf from zowe-cli imperative config
+            expect(apimlConfigs[zosmfInx].pluginName).toBe("@zowe/cli");
+
+            // we get both entries for endevor
+            expect(apimlConfigs[endvInx1].gatewayUrl).toBe("endv_api/v2");
+            expect(apimlConfigs[endvInx2].gatewayUrl).toBe("endv_api/v1");
+
+            // we infer the jclcheck profile, which was not specified by that plugin
+            expect(apimlConfigs[jckInx].pluginName).toBe("@broadcom/jclcheck-for-zowe-cli");
+            expect(apimlConfigs[jckInx].connProfType).toBe("jclcheck");
         });
     });
 
@@ -65,7 +123,7 @@ describe("APIML Services unit tests", () => {
     });
 
     describe("convertApimlProfileInfoToProfileConfig", () => {
-        const temp: IApimlProfileInfo[] = [
+        const temp: any[] = [ //TODO: fix compile errors: IApimlProfileInfo [] = [
             {
                 profName: "test1",
                 profType: "type1",
